# artifacts/api-server/src/routes/student.ts
import { Router, type IRouter } from "express";
import { z } from "zod";
import {
  db,
  studentsTable,
  classesTable,
  avatarConfigSchema,
} from "@workspace/db";
import { eq, and } from "drizzle-orm";
import {
  GetStudentProfileResponse,
  UpdateStudentAvatarBody,
  UpdateStudentAvatarResponse,
} from "@workspace/api-zod";
import { isAccessoriesUnlocked, isPetUnlocked } from "../lib/avatarUnlocks";
import { logActivity } from "../lib/activity-logs";

const router: IRouter = Router();

const UpdateStudentNameBody = z.object({
  name: z.string().min(1).max(120),
});

export async function getOrCreateStudent(userId: string) {
  const existing = await db.query.studentsTable.findFirst({
    where: eq(studentsTable.replitUserId, userId),
  });
  if (existing) return existing;

  // Fallback: the auth callback normally creates the row, but if a route is
  // called before that happens, provision a minimal row from the session id.
  const [created] = await db
    .insert(studentsTable)
    .values({
      replitUserId: userId,
      name: "صديقنا البطل",
      role: "student",
      nameConfirmed: false,
      points: 0,
    })
    .returning();
  return created;
}

async function enrichStudentProfile(student: typeof studentsTable.$inferSelect) {
  let className: string | null = null;
  let teacherName: string | null = null;
  let teacherEmail: string | null = null;

  if (student.classId) {
    const cls = await db.query.classesTable.findFirst({
      where: eq(classesTable.id, student.classId),
    });
    if (cls) {
      className = cls.name;
      if (cls.teacherId) {
        const teacher = await db.query.studentsTable.findFirst({
          where: eq(studentsTable.id, cls.teacherId),
        });
        if (teacher) {
          teacherName = teacher.name;
          teacherEmail = teacher.email;
        }
      }
    }
  }

  return {
    id: student.id,
    name: student.name,
    points: student.points,
    avatarConfig: avatarConfigSchema.parse(student.avatarConfig),
    classId: student.classId,
    className,
    teacherName,
    teacherEmail,
    imageUrl: student.imageUrl,
  };
}

router.get("/student/me", async (req, res) => {
  if (!req.isAuthenticated()) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  const userId = req.user.id;

  const student = await getOrCreateStudent(userId);
  const profile = await enrichStudentProfile(student);
  const data = GetStudentProfileResponse.parse(profile);
  res.json(data);
});

router.patch("/student/me", async (req, res) => {
  if (!req.isAuthenticated()) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  const userId = req.user.id;

  const body = UpdateStudentNameBody.parse(req.body);
  const student = await getOrCreateStudent(userId);

  const [updated] = await db
    .update(studentsTable)
    .set({ name: body.name, nameConfirmed: true })
    .where(eq(studentsTable.id, student.id))
    .returning();

  await logActivity(student.id, {
    type: "name_change",
    title: "غيّر الاسم",
    description: `تم تحديث الاسم إلى "${body.name}"`,
  });

  const profile = await enrichStudentProfile(updated);
  const data = GetStudentProfileResponse.parse(profile);
  res.json(data);
});

router.delete("/student/me", async (req, res) => {
  if (!req.isAuthenticated()) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  const userId = req.user.id;

  const student = await getOrCreateStudent(userId);
  await db.delete(studentsTable).where(eq(studentsTable.id, student.id));
  res.json({ deleted: true });
});

router.patch("/student/avatar", async (req, res) => {
  if (!req.isAuthenticated()) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  const userId = req.user.id;

  const body = UpdateStudentAvatarBody.parse(req.body);
  const student = await getOrCreateStudent(userId);

  if (!isAccessoriesUnlocked(body.accessories, student.points)) {
    res.status(400).json({ error: "هذا الإكسسوار غير مفتوح بعد" });
    return;
  }
  if (!isPetUnlocked(body.pet, student.points)) {
    res.status(400).json({ error: "هذا الحيوان الأليف غير مفتوح بعد" });
    return;
  }

  const [updated] = await db
    .update(studentsTable)
    .set({ avatarConfig: body })
    .where(eq(studentsTable.id, student.id))
    .returning();

  await logActivity(student.id, {
    type: "avatar_change",
    title: "غيّر الشخصية الكرتونية",
    description: "تم تحديث مظهر الشخصية",
    metadata: JSON.stringify({ gender: body.gender }),
  });

  const profile = await enrichStudentProfile(updated);
  const data = UpdateStudentAvatarResponse.parse(profile);
  res.json(data);
});

export default router;