# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /student/me:
    get:
      operationId: getStudentProfile
      tags: [student]
      summary: Get the signed-in student's profile
      description: Returns the current student's profile, creating one on first visit (JIT-provisioned from the Clerk account).
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "401":
          description: Not signed in
    patch:
      operationId: updateStudentName
      tags: [student]
      summary: Capture the user's full name on first sign-in
      description: Sets the in-platform name and marks the onboarding name step as complete.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameCaptureBody"
      responses:
        "200":
          description: Updated student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "401":
          description: Not signed in
    delete:
      operationId: deleteStudentAccount
      tags: [student]
      summary: Delete the signed-in student's account
      description: Deletes the Clerk user and the local student row. This action is irreversible.
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                required: [deleted]
        "401":
          description: Not signed in
        "500":
          description: Deletion failed
  /student/avatar:
    patch:
      operationId: updateStudentAvatar
      tags: [student]
      summary: Update the signed-in student's avatar customization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvatarConfig"
      responses:
        "200":
          description: Updated student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "401":
          description: Not signed in
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves object entities uploaded via presigned URLs.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir.
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    AvatarConfig:
      type: object
      properties:
        bgColor:
          type: string
        accessories:
          type: array
          items:
            type: string
        gender:
          type: string
          enum: [male, female]
        pet:
          type: string
        nickname:
          type: string
          maxLength: 30
          default: ""
        frame:
          type: string
          default: "none"
        badges:
          type: array
          items:
            type: string
          default: []
      required:
        - bgColor
        - accessories
        - gender
        - pet
        - nickname
        - frame
        - badges
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        points:
          type: integer
        avatarConfig:
          $ref: "#/components/schemas/AvatarConfig"
        classId:
          oneOf:
            - type: integer
            - type: "null"
        className:
          oneOf:
            - type: string
            - type: "null"
        teacherName:
          oneOf:
            - type: string
            - type: "null"
        teacherEmail:
          oneOf:
            - type: string
            - type: "null"
        imageUrl:
          oneOf:
            - type: string
            - type: "null"
      required:
        - id
        - name
        - points
        - avatarConfig
        - classId
        - className
        - teacherName
        - teacherEmail
        - imageUrl
    NameCaptureBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
      required:
        - name
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
...